Split out logger helpers into own file.
